tools/requirements.txt
aiohttp>=3.9.0
boto3==1.42.24
python-dotenv>=1.0.0
warrant-lite==1.0.4


